requirements.txt
# TrOCR and transformers
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer vision and image processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Evaluation metrics
jiwer>=3.0.0
rapidfuzz>=3.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Progress bars and logging
tqdm>=4.65.0
wandb>=0.15.0

# Configuration
PyYAML>=6.0
omegaconf>=2.3.0

# Jupyter and development
jupyter>=1.0.0
ipywidgets>=8.0.0
jupyterlab>=4.0.0

# XML processing (for ALTO integration)
lxml>=4.9.0
beautifulsoup4>=4.12.0

# OCR evaluation
editdistance>=0.6.0
Levenshtein>=0.21.0

# Datasets and data loading
datasets>=2.14.0
accelerate>=0.24.0

# Optional: for advanced training
deepspeed>=0.10.0
lightning>=2.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
